Split basic node file to account for Python 3 meta class changes
